# api/openapi.yaml
openapi: 3.0.3
info:
  title: Book Store API
  description: |
    Complete REST API specification for the Book Store eCommerce application.
    
    ## Authentication
    Protected endpoints require a JWT Bearer token obtained from `/api/v1/auth/login` or `/api/v1/auth/register`.
    Include it in the request header: `Authorization: Bearer <token>`
    
    ## Phase 1 Notes
    - Payment processing uses a **mock implementation**. No real transactions are made.
    - Password reset tokens are returned directly in the response (no email service in Phase 1).
    - All prices are in **Indian Rupees (INR / ₹)**.
  version: "1.0"
  contact:
    name: Book Store API Team
servers:
  - url: http://localhost:8080/api/v1
    description: Local development server
  - url: https://api.bookstore.com/api/v1
    description: Production server
...
security: []
paths:
...
  /home/recommended:
    get:
      tags: [Home]
      summary: Get Recommended for You books
      operationId: getRecommended
      parameters:
        - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: Recommended books
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/BookSummary'
  /home/bestsellers:
    get:
      tags: [Home]
      summary: Get Bestsellers this Month
      operationId: getBestsellers
      parameters:
        - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: Bestseller books
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/BookSummary'
  /home/new-launches:
    get:
      tags: [Home]
      summary: Get New Launches
      operationId: getNewLaunches
      parameters:
        - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: New launch books
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/BookSummary'

  # ─────────────────────────────────────────────────
  # GENRES
  # ─────────────────────────────────────────────────
...
  /payments/process:
    post:
      tags: [Payments]
      summary: Process payment for a placed order (mock implementation)
      operationId: processPayment
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PaymentRequest'
      responses:
        '200':
          description: Payment processed successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/PaymentResult'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
  /payments/{orderId}:
    get:
      tags: [Payments]
      summary: Get payment record for a specific order
      operationId: getPaymentByOrderId
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/OrderId'
      responses:
        '200':
          description: Payment record
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/PaymentRecord'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

  # ─────────────────────────────────────────────────
  # ORDERS
  # ─────────────────────────────────────────────────
...
  /orders/{orderId}:
    get:
      tags: [Orders]
      summary: Get full detail of a single order
      operationId: getOrderById
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/OrderId'
      responses:
        '200':
          description: Order detail
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/OrderDetail'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /orders/{orderId}/buy-again:
    post:
      tags: [Orders]
      summary: Re-add all items from a past order to the current cart
      operationId: buyAgain
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/OrderId'
      responses:
        '200':
          description: Items added to cart
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/BuyAgainResult'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

# ─────────────────────────────────────────────────────────────────────────────
# COMPONENTS
# ─────────────────────────────────────────────────────────────────────────────
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access token obtained from /auth/login or /auth/register

  # ── PARAMETERS ────────────────────────────────────────────────────────────
  parameters:
...
    OrderId:
      name: orderId
      in: path
      required: true
      description: Unique order identifier
      schema:
        type: integer
        format: int64
        example: 7
...
    Limit:
      name: limit
      in: query
      description: Maximum number of items to return
      schema:
        type: integer
        default: 10
        minimum: 1
        maximum: 20

  # ── RESPONSES ─────────────────────────────────────────────────────────────
  responses:
    BadRequest:
      description: Validation failed or malformed request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            status: 400
            error: BAD_REQUEST
            message: Validation failed
            errors:
              - field: email
                message: must be a valid email address
            timestamp: "2025-07-21T10:30:00Z"
            path: /api/v1/auth/register
    Unauthorized:
      description: Missing or invalid JWT token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            status: 401
            error: UNAUTHORIZED
            message: JWT token is missing or invalid
            timestamp: "2025-07-21T10:30:00Z"
            path: /api/v1/cart
    Forbidden:
      description: Valid token but insufficient permissions
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: Requested resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            status: 404
            error: NOT_FOUND
            message: Book with id 999 not found
            timestamp: "2025-07-21T10:30:00Z"
            path: /api/v1/books/999
...
    UnprocessableEntity:
      description: Business rule violation
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            status: 422
            error: UNPROCESSABLE_ENTITY
            message: Coupon SAVE100 has expired
            timestamp: "2025-07-21T10:30:00Z"
            path: /api/v1/checkout/validate-coupon

  # ── SCHEMAS ───────────────────────────────────────────────────────────────
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          type: object
          nullable: true
        message:
          type: string
          example: Operation successful
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        status:
          type: integer
          example: 400
        error:
          type: string
          example: BAD_REQUEST
        message:
          type: string
          example: Validation failed
        errors:
          type: array
          nullable: true
          items:
            type: object
            properties:
              field:
                type: string
              message:
                type: string
        timestamp:
          type: string
          format: date-time
        path:
          type: string
...
    Genre:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 3
        name:
          type: string
          example: Science Fiction
        slug:
          type: string
          example: science-fiction

    # ── Book ────────────────────────────────────────────────────────────
    AuthorSummary:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 3
        name:
          type: string
          example: Daniel Reed
    BookSummary:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 12
        title:
          type: string
          example: The Joy of Minimalism
        author:
          $ref: '#/components/schemas/AuthorSummary'
        coverImageUrl:
          type: string
          nullable: true
          example: https://cdn.bookstore.com/covers/joy-minimalism.jpg
        shortDescription:
          type: string
          nullable: true
          example: Declutter your life to uncover peace, clarity, and joy.
        format:
          type: string
          enum: [PAPERBACK, HARDCOVER, EBOOK]
          example: PAPERBACK
        genres:
          type: array
          items:
            $ref: '#/components/schemas/Genre'
        price:
          type: number
          format: double
          example: 149.00
        estimatedDeliveryDate:
          type: string
          format: date
          example: "2025-07-28"
        averageRating:
          type: number
          format: double
          nullable: true
          example: 4.5
        copiesSold:
          type: integer
          example: 145
...
    CartItem:
      type: object
      properties:
        cartItemId:
          type: integer
          format: int64
          example: 22
        bookId:
          type: integer
          format: int64
          example: 12
        title:
          type: string
          example: The Joy of Minimalism
        coverImageUrl:
          type: string
          nullable: true
        author:
          $ref: '#/components/schemas/AuthorSummary'
        format:
          type: string
          enum: [PAPERBACK, HARDCOVER, EBOOK]
        genres:
          type: array
          items:
            $ref: '#/components/schemas/Genre'
        unitPrice:
          type: number
          format: double
          example: 149.00
        quantity:
          type: integer
          minimum: 1
          example: 1
        lineTotal:
          type: number
          format: double
          example: 149.00
        estimatedDeliveryDate:
          type: string
          format: date
...
    OrderItem:
      type: object
      properties:
        id:
          type: integer
          format: int64
        bookId:
          type: integer
          format: int64
        title:
          type: string
          example: The Joy of Minimalism
        coverImageUrl:
          type: string
          nullable: true
        format:
          type: string
          enum: [PAPERBACK, HARDCOVER, EBOOK]
        unitPrice:
          type: number
          format: double
          example: 149.00
        quantity:
          type: integer
          example: 1
        lineTotal:
          type: number
          format: double
          example: 149.00
...
    CardDetails:
      type: object
      required: [cardNumber, nameOnCard, cvv, expiryDate]
      properties:
        cardNumber:
          type: string
          pattern: '^\d{13,19}$'
          example: "4111111111111111"
        nameOnCard:
          type: string
          example: Arjun Patel
        cvv:
          type: string
          pattern: '^\d{3,4}$'
          example: "123"
        expiryDate:
          type: string
          pattern: '^(0[1-9]|1[0-2])\/\d{4}$'
          example: "12/2027"
    UpiDetails:
      type: object
      required: [upiId]
      properties:
        upiId:
          type: string
          example: arjun@okaxis
    PaymentRequest:
      type: object
      required: [orderId, paymentMethod]
      properties:
        orderId:
          type: integer
          format: int64
          example: 7
        paymentMethod:
          type: string
          enum: [CREDIT_CARD, DEBIT_CARD, UPI, WALLET]
          example: CREDIT_CARD
        cardDetails:
          $ref: '#/components/schemas/CardDetails'
          description: Required for CREDIT_CARD and DEBIT_CARD
        upiDetails:
          $ref: '#/components/schemas/UpiDetails'
          description: Required for UPI
    PaymentRecord:
      type: object
      properties:
        paymentId:
          type: integer
          format: int64
          example: 4
        orderId:
          type: integer
          format: int64
          example: 7
        paymentMethod:
          type: string
          enum: [CREDIT_CARD, DEBIT_CARD, UPI, WALLET]
        payableAmount:
          type: number
          format: double
          example: 469.00
        status:
          type: string
          enum: [PENDING, SUCCESS, FAILED]
          example: SUCCESS
        paidAt:
          type: string
          format: date-time
          nullable: true
        createdAt:
          type: string
          format: date-time
    PaymentResult:
      allOf:
        - $ref: '#/components/schemas/PaymentRecord'
        - type: object
          properties:
            order:
              $ref: '#/components/schemas/OrderDetail'

    # ── Order ────────────────────────────────────────────────────────────
    OrderSummary:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 7
        orderNumber:
          type: string
          example: BST-20250721-00007
        status:
          type: string
          enum: [PLACED, CONFIRMED, SHIPPED, DELIVERED, CANCELLED]
          example: CONFIRMED
        totalAmount:
          type: number
          format: double
          example: 469.00
        itemCount:
          type: integer
          example: 2
        placedAt:
          type: string
          format: date-time
        estimatedDeliveryDate:
          type: string
          format: date
        items:
          type: array
          items:
            $ref: '#/components/schemas/OrderItem'
    DeliveryAddressSnapshot:
      type: object
      properties:
        firstName:
          type: string
        lastName:
          type: string
        addressLine1:
          type: string
        addressLine2:
          type: string
          nullable: true
        city:
          type: string
        state:
          type: string
        country:
          type: string
        pinCode:
          type: string
    OrderDetail:
      allOf:
        - $ref: '#/components/schemas/OrderSummary'
        - type: object
          properties:
            subtotal:
              type: number
              format: double
            taxAmount:
              type: number
              format: double
            deliveryCharge:
              type: number
              format: double
            discountAmount:
              type: number
              format: double
            couponCode:
              type: string
              nullable: true
            deliveryAddress:
              $ref: '#/components/schemas/DeliveryAddressSnapshot'
            payment:
              $ref: '#/components/schemas/PaymentRecord'
...
    BuyAgainResult:
      type: object
      properties:
        cartId:
          type: integer
          format: int64
        itemCount:
          type: integer
        itemsAdded:
          type: integer
        items:
          type: array
          items:
            $ref: '#/components/schemas/CartItem'